poller/requirements.txt
pika
redis
psycopg2-binary